Replaced the build-script git clone with a local copy.
Previously, the Vagrant puppet manifest would clone the build scripts
from GitHub during the server build process.  This meant that if you
edited the configuration on the host machine, you'd still be getting
the un-edited version from GitHub.

Now, Vagrant mounts the host's copy of the built scripts on the
virtual machine, and then copies them for use in building the server.
This means any local edits to the build config will be immediately
available to the virtual machine.